+++ packages/auth/src/tests/activate-account-flow.cy-spec.tsx
@@ -17,6 +17,7 @@ const defaultAuthPlugin = {
     loginUrl: '/account/login',
     logoutUrl: '/account/logout',
     activateUrl: '/account/activate',
+    acceptInvitationUrl: '/account/invitation/accept',
     forgetPasswordUrl: '/account/forget-password',
     resetPasswordUrl: '/account/reset-password',
   },